requirements.txt
beautifulsoup4~=4.9.3
bs4~=0.0.1
nekos.py
python-decouple
Pillow~=8.3.1
Pyrogram~=1.2.9
requests~=2.26.0
TgCrypto
wget~=3.2
